Update the function to split the sections
I did the regular expression based on the pattern for each section according to MAS narrative. However, it might not be the case for other sectors, the function should be customized by sector. Also, further adjustment needed on identifying the table and the indicators (comparatively easier, because the table items can be identified easily with the summary_tagging.